cpu/esp32: fixes frame size handling in esp_wifi
The size of received and transmitted frames was stored in an uint8_t, which did not allow to process frames larger than 255 octets. However, WiFi has an MTU of 1500 octets.

+++ cpu/esp32/esp-wifi/esp_wifi_netdev.c
@@ -298,7 +298,7 @@ static int _esp_wifi_recv(netdev_t *netdev, void *buf, size_t len, void *info)
 
     mutex_lock(&dev->dev_lock);
 
-    uint8_t size = dev->rx_len;
+    uint16_t size = dev->rx_len;
 
     if (!buf && !len) {
         /* return the size without dropping received data */
